modularised testing for clarity
